backend/requirements.txt
Django==4.2.25
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
pymongo==4.10.1
PyJWT==2.9.0
